+++ osori-cave-ui/src/pages/navigation/index.js
@@ -2,5 +2,14 @@ import Navigation from './Navigation'
 import AddChildrenForm from './AddChildrenForm'
 import ModifyForm from './ModifyForm'
 import NavigationModal from './NavigationModal'
+import NavigationTree from './NavigationTree'
+import NavigationList from './NavigationList'
 
-export {Navigation, AddChildrenForm, NavigationModal, ModifyForm}
+export {
+    Navigation,
+    AddChildrenForm,
+    NavigationModal,
+    ModifyForm,
+    NavigationTree,
+    NavigationList
+}
